refactor: move mp_hal_stdin_(rx|tx)_chr from uart-qemu.c to mphalport.c
