+++ raspberrypi/mphalport.h
@@ -7,3 +7,6 @@ mp_uint_t mp_hal_ticks_us(void);
 mp_uint_t mp_hal_ticks_ms(void);
 
 void mp_hal_set_interrupt_char(int c);
+int mp_hal_stdin_rx_chr(void);
+void mp_hal_stdout_tx_strn(const char *str, mp_uint_t len);
+
